Fixed that cysimdjson was set to only "21.11b2"
I've encountered an issue when trying to install bspump through pip on Mac. Whenever I tried to install it, it would fail on cysimdjson, however when I tried to install cysimdjson on it's own it worked.

As it turns out cysimdjson version 21.11b2 (The version that bspump currently depends on) is not installable on Mac for some reason. But cysimdjson version 21.11 is.

I don't know if there's a reason for using only "21.11b2" but if there's not it should probably be latest.